move remaining pomodoro functions to focus component
