+++ app/components/home/Focus.jsx
@@ -1,16 +1,42 @@
+import { usePomodoroMode } from "@/app/hooks/usePomodoroMode";
 import { pomodoroSelectors } from "@/app/redux/selectors/pomodoroSelectors";
-import { stopPomodoro, togglePomodoro } from "@/app/redux/slices/pomodoroSlice";
+import { settingsSelectors } from "@/app/redux/selectors/settingsSelectors";
+import {
+  stopPomodoro,
+  timeTick,
+  togglePomodoro,
+  updateCount,
+  updateMode,
+  updateProgress,
+  updateTimeLeft,
+} from "@/app/redux/slices/pomodoroSlice";
 import { Pause, Play, SkipForward, RefreshCcw } from "lucide-react";
+import { useCallback, useEffect, useMemo, useRef } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-const Focus = ({ formatTime, skipPomodoro, totalTime, currentTask }) => {
+const Focus = ({ formatTime, currentTask }) => {
   const dispatch = useDispatch();
+  const completedRef = useRef(false);
+  const intervalRef = useRef(null);
+  const { getMode, getModeTime } = usePomodoroMode();
 
-  const { progress, currentMode } = useSelector(pomodoroSelectors);
+  const { progress, currentMode, pomodoroCount } =
+    useSelector(pomodoroSelectors);
+  const { autoStartNext, focusTime, shortBreakTime, longBreakTime } =
+    useSelector(settingsSelectors);
+
   const isRunning = useSelector((state) => state.pomodoro.isRunning);
   const timeLeft = useSelector((state) => state.pomodoro.timeLeft);
   const circumference = 2 * Math.PI * 180;
   const strokeDashoffset = circumference * (1 - progress / 100);
+
+  const totalTime = useMemo(() => {
+    return {
+      focusTime: focusTime * 60,
+      shortBreakTime: shortBreakTime * 60,
+      longBreakTime: longBreakTime * 60,
+    }[currentMode];
+  }, [focusTime, shortBreakTime, longBreakTime, currentMode]);
 
   const runPomodoro = () => {
     dispatch(togglePomodoro());
@@ -19,6 +45,81 @@ const Focus = ({ formatTime, skipPomodoro, totalTime, currentTask }) => {
     dispatch(stopPomodoro());
     dispatch(updateTimeLeft({ time: totalTime }));
   };
+  const skipPomodoro = () => {
+    dispatch(stopPomodoro());
+    const modeTime = getModeTime();
+    if (currentMode === "focusTime") {
+      dispatch(updateCount());
+      const mode = getMode();
+      dispatch(updateMode({ mode: mode }));
+      dispatch(updateTimeLeft({ time: modeTime }));
+    } else {
+      dispatch(updateMode({ mode: "focusTime" }));
+      dispatch(updateTimeLeft({ time: modeTime }));
+    }
+  };
+
+  useEffect(() => {
+    if (isRunning) {
+      completedRef.current = false;
+    }
+  }, [isRunning]);
+
+  const handlePomodoroComplete = useCallback(() => {
+    if (completedRef.current) return;
+    completedRef.current = true;
+    dispatch(stopPomodoro());
+    const modeTime = getModeTime();
+    if (currentMode === "focusTime") {
+      dispatch(updateCount());
+      // ! ADD THIS TO REDUX
+      // setStreak((prev) => {
+      //   const streak = prev + 1;
+      //   localStorage.setItem("streak", JSON.stringify(streak));
+      //   return streak;
+      // });
+
+      const mode = getMode();
+      dispatch(updateMode({ mode: mode }));
+      dispatch(updateTimeLeft({ time: modeTime }));
+    } else {
+      dispatch(updateMode({ mode: "focusTime" }));
+      dispatch(updateTimeLeft({ time: modeTime }));
+    }
+    // ! note: enable user add aut-start time & add countdown before autostart
+    if (autoStartNext) {
+      setTimeout(() => dispatch(togglePomodoro()), 3000);
+    }
+  }, [dispatch, currentMode, autoStartNext, pomodoroCount]);
+
+  useEffect(() => {
+    if (!isRunning) return;
+    intervalRef.current = setInterval(() => {
+      dispatch(timeTick());
+      if (timeLeft <= 1) {
+        handlePomodoroComplete();
+      }
+    }, 1000);
+
+    return () => clearInterval(intervalRef.current);
+  }, [dispatch, timeLeft, isRunning, handlePomodoroComplete]);
+  const times = useMemo(
+    () => ({
+      focusTime: focusTime * 60,
+      shortBreakTime: shortBreakTime * 60,
+      longBreakTime: longBreakTime * 60,
+    }),
+    [focusTime, shortBreakTime, longBreakTime]
+  );
+  useEffect(() => {
+    dispatch(updateTimeLeft({ time: times[currentMode] }));
+  }, [dispatch, times, currentMode]);
+
+  useEffect(() => {
+    if (!totalTime) return;
+    const progressT = ((totalTime - timeLeft) / totalTime) * 100;
+    dispatch(updateProgress({ time: progressT }));
+  }, [dispatch, totalTime, timeLeft]);
   return (
     <div className="timer-container">
       <div className="timer-circle">
